Make reviewers required changes
